refactor: extract test class from CoverageGapTests
